+++ lib/sapi_client.rb
@@ -13,6 +13,7 @@ require 'sapi_client/application'
 require 'sapi_client/instance'
 require 'sapi_client/view'
 require 'sapi_client/sapi_resource'
+require 'sapi_client/endpoint_values'
 
 module SapiClient
   class Error < StandardError; end